Stealing is bad (#33)
* Added consequences for stealing

Added consequences for letting your stuff get stolen. as well as fixing the timer so it stared and stops based on quest system. Added theme to timer and quest finished text.

* Remove no longer needed start area/win area and clean up names

Removed the start area and win area for the timer, as timer is set by quest system now. Also named the pit cube, and hub background nodes.

------